fix(query_keeper_test):Update query keepers test logic for multiple modules
The logic of the query keeper tests for the modules "assets", "classifications", "identities", "metas", and "splits" have been updated. This includes changes to the import statements, changes to Initialize and Enquire functions, as well as the addition of the "context" package. The mappable prototype is replaced with appropriate record or mapper prototype and handling errors where necessary.